requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
celery
redis
drf-yasg